+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
-## 1.0.3
-* Fixed the issue of inaccurate marker position.
-* Limit screenshot resolution to 1080p.
+## 1.2.0
+* feat: Support using change marker button to view screenshot.
+* fix: Multiple nearby markers cannot be selected accurately.
+
+## 1.1.0
+* feat: Add a flash effect when taking screenshot.
+* feat: Compatible with RenoDX and Luma-Framework. Thanks alexkiri.
 
 ## 1.0.4
-* Fixed the issue that screenshots cannot be taken in the new game.
+* fix: Fixed the issue that screenshots cannot be taken in the new game.
+
+## 1.0.3
+* feat: Limit screenshot resolution to 1080p.
+* fix: Fixed the issue of inaccurate marker position.